+++ src/main/kotlin/at.phatbl.shellexec/ShellCommand.kt
@@ -15,38 +15,84 @@ data class ShellCommand(
         // 20m
         private const val defaultTimeout: Long = 1200
 
-        private const val uninitializedExitValue = -999
+        // Used to track whether we've received an updated exit code from the process.
+        const val uninitializedExitValue = -999
 
         private const val bufferSize = 2 * 1024 * 1024
     }
 
-    lateinit var process: Process
-
-    var standardOutput: OutputStream? = null
-    var errorOutput: OutputStream? = null
-//    val standardInput: InputStream
-
-    val stdout: String
-        get() = stream2String(process.inputStream)
-
-    val stderr: String
-        get() = stream2String(process.errorStream)
-
-    var exitValue: Int = uninitializedExitValue
-
-    val succeeded: Boolean
-        get() = exitValue == 0
-
-    val failed: Boolean
-        get() = !succeeded
-
-    var readLimit = bufferSize
-
-    /** Time allowed for command to run. Defaults to 20m */
+    /**
+     * Time allowed for command to run. Defaults to 20m
+     */
     var timeout = defaultTimeout
 
     /**
+     * Exposes the exit code of the underlying process. Defaults to -999 until the process has exited.
+     */
+    var exitValue: Int = uninitializedExitValue
+
+    /**
+     * True if the process ended with a successful exit code.
+     */
+    val succeeded: Boolean
+        get() = exitValue == 0
+
+    /**
+     * False if the process did not return a successful exit code.
+     */
+    val failed: Boolean
+        get() = !succeeded
+
+    /**
+     * Exposes a stream of command output from the underlying process if you provide an OutputStream.
+     * Mutually exclusive with stdout. Defaults to null.
+     * @see stdout
+     */
+    var standardOutput: OutputStream? = null
+
+    /**
+     * Exposes a stream of command error output from the underlying process if you provide an OutputStream.
+     * Mutually exclusive with stderr. Defaults to null.
+     * @see stderr
+     */
+    var errorOutput: OutputStream? = null
+
+    /**
+     * Convenience property for accessing stdout from command as a string after the command has finished running.
+     * Will always be null if you provide an OutputStream to standardOutput.
+     * @see standardOutput
+     */
+    val stdout: String?
+        get() {
+            val output = outputFile ?: return null
+            return output.bufferedReader().use { it.readText() }
+        }
+
+    /**
+     * Convenience property for accessing stderr from command as a string after the command has finished running.
+     * Will always be null if you provide an OutputStream to errorOutput.
+     * @see errorOutput
+     */
+    val stderr: String?
+        get() {
+            val errors = errorFile ?: return null
+            return errors.bufferedReader().use { it.readText() }
+        }
+
+    /**
+     * File containing stdout from command. null if an OutputStream is provided to standardOutput before start is called.
+     */
+    private var outputFile: File? = null
+
+    /**
+     * File containing stderr from command. null if an OutputStream is provided to errorOutput before start is called.
+     */
+    private var errorFile: File? = null
+
+    /**
      * Runs the command.
+     *
+     * @throws ShellCommandTimeoutException if the command process did not finish before the timeout.
      */
     fun start() {
         baseDir.mkdirs()
@@ -54,77 +100,55 @@ data class ShellCommand(
         val pb = ProcessBuilder("bash", "-c", command)
             .directory(baseDir)
 
+        val outputStream = standardOutput
+        val errorStream = errorOutput
+
+        if (outputStream == null) {
+            outputFile = createTempFile("shellexec-", "-output.log")
+            pb.redirectOutput(outputFile)
+        }
+        if (errorStream == null) {
+            errorFile = createTempFile("shellexec-", "-error.log")
+            pb.redirectError(errorFile)
+        }
+
         // Launch the process
-        process = pb.start()
+        val process = pb.start()
 
-        process.inputStream.mark(readLimit)
-        process.errorStream.mark(readLimit)
-
-        if (standardOutput != null) {
-            copy(input = process.inputStream, output = standardOutput!!)
+        if (outputStream != null) {
+            copy(input = process.inputStream, output = outputStream)
         }
-        if (errorOutput != null) {
-            copy(input = process.errorStream, output = errorOutput!!)
-        }
-
-        try {
-            process.inputStream.reset()
-        } catch (e: Exception) {
-            val stdout = standardOutput
-            if (stdout != null) {
-                val errorMessage = "Could not reset input stream. Mark with Readlimit $readLimit not found.".toByteArray()
-                stdout.write(errorMessage)
-            }
-        }
-
-        try {
-            process.errorStream.reset()
-        } catch (e: Exception) {
-            val stderr = errorOutput
-            if (stderr != null) {
-                val errorMessage = "Could not reset error stream. Mark with Readlimit $readLimit not found.".toByteArray()
-                stderr.write(errorMessage)
-            }
+        if (errorStream != null) {
+            copy(input = process.errorStream, output = errorStream)
         }
 
         try {
             process.waitFor(timeout, TimeUnit.SECONDS)
-            exitValue = process.exitValue()
-        } catch (e: Exception) {
-            // Handle timeouts
-            println("ShellCommand timeout: $e")
+
+            // Check to see if the process has quit. Otherwise calling exitValue throws IllegalThreadStateException
+            if (!process.isAlive) {
+                exitValue = process.exitValue()
+            }
+        } catch (e: InterruptedException) {
+            val message = "Command timeout, exceeded $timeout second limit."
+            throw ShellCommandTimeoutException(message, e)
         }
     }
 
     /**
-     * Utility function which converts an input stream into a string.
+     * Passes characters from the input stream to the output stream.
      */
-    private fun stream2String(stream: InputStream): String {
-        return try {
-            val reader = BufferedReader(InputStreamReader(stream))
-            val builder = StringBuilder()
-            val lineSeparator = System.getProperty("line.separator")
-            reader.forEachLine { line ->
-                builder.append(line)
-                builder.append(lineSeparator)
-            }
-
-            builder.toString()
-        } catch (e: Exception) {
-            return ""
-        }
-    }
-
     @Throws(IOException::class)
     private fun copy(input: InputStream, output: OutputStream) {
-        try {
-            val buffer = ByteArray(bufferSize)
-            var bytesRead = input.read(buffer)
-            while (bytesRead != -1) {
-                output.write(buffer, 0, bytesRead)
-                bytesRead = input.read(buffer)
+        input.use {
+            output.use {
+                val buffer = ByteArray(bufferSize)
+                var bytesRead = input.read(buffer)
+                while (bytesRead != -1) {
+                    output.write(buffer, 0, bytesRead)
+                    bytesRead = input.read(buffer)
+                }
             }
-            //If needed, close streams.
-        } finally { }
+        }
     }
 }