+++ src/test/kotlin/at.phatbl.shellexec/ShellCommandSpek.kt
@@ -4,9 +4,11 @@ import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
 import org.junit.rules.TemporaryFolder
+import java.io.ByteArrayOutputStream
 import java.io.File
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
+import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
 object ShellCommandSpek : Spek({
@@ -16,7 +18,7 @@ object ShellCommandSpek : Spek({
         beforeEachTest {}
 
         it("can run a simple command") {
-            shellCommand = ShellCommand(baseDir = File("."), command = "true")
+            shellCommand = ShellCommand(command = "true")
             shellCommand.start()
 
             assertTrue(shellCommand.succeeded)
@@ -25,7 +27,7 @@ object ShellCommandSpek : Spek({
         }
 
         it("can run a failing command") {
-            shellCommand = ShellCommand(baseDir = File("."), command = "false")
+            shellCommand = ShellCommand(command = "false")
             shellCommand.start()
 
             assertTrue(shellCommand.failed)
@@ -34,19 +36,63 @@ object ShellCommandSpek : Spek({
         }
 
         it("can generate standard output") {
-            shellCommand = ShellCommand(baseDir = File("."), command = "echo Hello World!")
+            shellCommand = ShellCommand(command = "echo Hello World!")
             shellCommand.start()
 
             assertTrue(shellCommand.succeeded)
             assertEquals("Hello World!\n", shellCommand.stdout)
+            assertNull(shellCommand.standardOutput)
+        }
+
+        it("has empty but not null standard output") {
+            shellCommand = ShellCommand(command = "echo -n")
+            shellCommand.start()
+
+            assertTrue(shellCommand.succeeded)
+            assertEquals("", shellCommand.stdout)
+            assertNull(shellCommand.standardOutput)
+        }
+
+        it("can generate standard output to a stream") {
+            shellCommand = ShellCommand(command = "echo Hello World!")
+            val stream = ByteArrayOutputStream()
+            shellCommand.standardOutput = stream
+
+            shellCommand.start()
+
+            assertTrue(shellCommand.succeeded)
+            assertEquals("Hello World!\n", String(stream.toByteArray()))
+            assertNull(shellCommand.stdout)
         }
 
         it("can generate error output") {
-            shellCommand = ShellCommand(baseDir = File("."), command = "echo This is an error. >&2")
+            shellCommand = ShellCommand(command = "echo This is an error. >&2")
             shellCommand.start()
 
             assertTrue(shellCommand.succeeded)
             assertEquals("This is an error.\n", shellCommand.stderr)
+            assertNull(shellCommand.errorOutput)
+        }
+
+        it("has empty but not null error output") {
+            shellCommand = ShellCommand(command = "echo -n >&2")
+            shellCommand.start()
+
+            assertTrue(shellCommand.succeeded)
+            assertEquals("", shellCommand.stderr)
+            assertNull(shellCommand.errorOutput)
+        }
+
+        it("can generate error output to a stream") {
+            shellCommand = ShellCommand(command = "echo This is an error. >&2")
+            val stream = ByteArrayOutputStream()
+            shellCommand.errorOutput = stream
+
+            shellCommand.start()
+
+            assertTrue(shellCommand.succeeded)
+            assertEquals("This is an error.\n", String(stream.toByteArray()))
+            assertNull(shellCommand.stderr)
         }
 
         it("can invoke a command with spaces in the path") {
@@ -71,7 +117,7 @@ object ShellCommandSpek : Spek({
 
             assertTrue(shellCommand.succeeded)
             assertEquals("", stderr)
-            assertEquals("$fileContents\n", stdout)
+            assertEquals(fileContents, stdout)
 
             temporaryFolder.delete()
         }